+++ package.json
@@ -37,6 +37,7 @@
         "eslint": "^8.8.0",
         "jest": "^27.4.7",
         "node-notifier": "^10.0.0",
+        "safe-regex": "^2.1.1",
         "webpack": "^5.67.0",
         "webpack-cli": "^4.9.2",
         "webpack-node-externals": "^3.0.0"
