--- server/src/constants.json
@@ -1,14 +1,0 @@
-{
-    "MARKET_CAP_SIZES": {
-        "MEGA": "MEGA",
-        "LARGE": "LARGE",
-        "MID": "MID",
-        "SMALL": "SMALL",
-        "MICRO": "MICRO",
-        "NANO": "NANO"
-    },
-    "TRADE_UNIT_TYPES": {
-        "DOLLARS": "DOLLARS",
-        "SHARES": "SHARES"
-    }
-}
